% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{can_access_repo}
\alias{can_access_repo}
\title{Check Availability of TL::DR Repo}
\usage{
can_access_repo()
}
\value{
boolean
}
\description{
Check if we can access the repo (we have internet and the repo isn't down)
}
