% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display Page Documentation
Display a page, encoded as a length-one character vector}
\usage{
display(page)
}
\arguments{
\item{page}{}
}
\value{
a message to the user in the console or viewer pane depending on
  user options
}
\description{
Display Page Documentation
Display a page, encoded as a length-one character vector
}
