% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_html}
\alias{display_html}
\title{Display Page as HTML}
\usage{
display_html(page)
}
\arguments{
\item{page}{}
}
\value{
an HTML page to the Viewer pane
}
\description{
Display the page in the Viewer pane
}
