% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{dr}
\alias{dr}
\title{View a Quick Reference Page}
\usage{
dr(fun, namespace = NULL)
}
\arguments{
\item{fun}{the function}

\item{namespace}{the package the function is in}
}
\description{
download and display a simple quick reference page for a
  function
}
\examples{
\dontrun{
# pull up the quick-reference guide for the first function in dplyr
tl::dr(dplyr::first)

# you can also call it in one of these ways!
tl::dr(first, dplyr)
tl::dr("dplyr::first")
tl::dr("first", "dplyr")
}
}
