% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{format_text}
\alias{format_text}
\title{Format Text Output}
\usage{
format_text(page)
}
\arguments{
\item{page}{the text stream from the document page}
}
\value{
a text formatted version of the document page
}
\description{
Formats the document page output for nice display in the console
}
