% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_page}
\alias{get_page}
\title{Get Page}
\usage{
get_page(namespace, fun)
}
\arguments{
\item{namespace}{the package the function exists within}

\item{fun}{the function to return documenation for}
}
\value{
a page for presenation
}
\description{
Fetch the page from the online repository as a character vector of length one.
If the online repository cannot be reached the function returns from the local
install of the package.
}
