% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{notify_using_cache}
\alias{notify_using_cache}
\title{Using Cache}
\usage{
notify_using_cache()
}
\value{
prints a message
}
\description{
Advising the user that they're using a cached version of the documenation.
}
\details{
Let the user know we're using the cache
}
