% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_not_available.R
\name{page_not_available}
\alias{page_not_available}
\title{Check if Page exists}
\usage{
page_not_available(namespace, fun)
}
\arguments{
\item{namespace}{package the function exists within}

\item{fun}{function for documentation}
}
\value{
a message
}
\description{
If we don't have that page, tell the user and prompt them to submit one
}
