% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_page.R
\name{submit_page}
\alias{submit_page}
\title{Submit a Page to the Online Repository}
\usage{
submit_page(fun, namespace = NULL, location = getwd())
}
\arguments{
\item{fun}{name of function being documented}

\item{namespace}{the package the function is in}

\item{location}{the folder containing the package folder and markdown file
(see details).}
}
\description{
runs checks on the markdown file containing the tl::dr page, and
  provides instructions to submit the page to the github repository
}
\details{
the markdown file should be named <function_name.md>, where
  <function_name> is the name of the function, and be in a directory with the
  same name as the package. That package directory should be within the
  directory given by \code{location} (by default the current working
  directory). the easiest way to set this up is by using
  \code{\link{create_page}()}.
}
